refactor Binner to child of Slicer
